requirements.txt
annotated-types==0.6.0
anyio==4.3.0
click==8.1.7
exceptiongroup==1.2.1
fastapi==0.110.2
h11==0.14.0
idna==3.7
numpy==1.26.4
opencv-python==4.9.0.80
pillow==10.3.0
pydantic==2.7.1
pydantic_core==2.18.2
python-multipart==0.0.9
sniffio==1.3.1
starlette==0.37.2
typing_extensions==4.11.0
uvicorn==0.29.0
